streamlit_app/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "techshop-ui"
version = "0.1.0"
description = "Interfaz Streamlit para el agente TechShop — Curso LLMOps"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Curso LLMOps", email = "curso@example.com"},
]

dependencies = [
    "streamlit>=1.40.0",
    "techshop-agent",
    "langfuse>=3.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
techshop-ui = "app:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = ["app.py"]

[tool.uv]
package = false

[tool.uv.sources]
techshop-agent = {path = "..", editable = true}
